code/pip-requirements.txt
prometheus_client
requests
pycryptodome
Flask